{"schema":"decodex/commit/1","summary":"Modularize selection scrim affordances","authority":"manual"}
